Fixing cleanup.sh

+++ cleanup.sh
@@ -23,7 +23,7 @@ set -e
 if [ -z "$CHANNELS" ]; then
   OMAHA="$(curl -s https://omahaproxy.appspot.com/all.json)"
   if [ -z "$CHANNELS" ]; then
-    CHANNELS=$(jq -r '.[] | select(.os == "win64") | .versions[] | .channel' <<< "$OMAHA")
+    CHANNELS=$(jq -r '.[] | select(.os == "win64") | .versions[] | .channel' <<< "$OMAHA"|tr '\r\n' ' '|sed -e 's/ $//')
   fi
 fi
 if [ -z "$VERSIONS" ]; then
@@ -35,23 +35,25 @@ if [ -z "$VERSIONS" ]; then
   done
 fi
 
-pushd $SRC &> /dev/null
+echo "KEEP: ${CHANNELS[@]} ${VERSIONS[@]}"
 
 # cleanup old directories and archives
-pushd $SRC/out &> /dev/null
-DIRS=$(find . -maxdepth 1 -type d -printf "%f\n"|egrep '^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$'|egrep -v "($(join_by '|' $VERSIONS))"||:)
-if [ ! -z "$DIRS" ]; then
-  (set -x;
-    rm -rf $DIRS
-  )
+if [ -d $SRC/out ]; then
+  pushd $SRC/out &> /dev/null
+  DIRS=$(find . -maxdepth 1 -type d -printf "%f\n"|egrep '^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$'|egrep -v "($(join_by '|' $VERSIONS))"||:)
+  if [ ! -z "$DIRS" ]; then
+    (set -x;
+      rm -rf $DIRS
+    )
+  fi
+  ARCHIVES=$(find . -maxdepth 1 -type f -printf "%f\n"|egrep '^headless-shell-[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+\.tar\.bz2'|egrep -v "($(join_by '|' $VERSIONS))"||:)
+  if [ ! -z "$ARCHIVES" ]; then
+    (set -x;
+      rm -rf $ARCHIVES
+    )
+  fi
+  popd &> /dev/null
 fi
-ARCHIVES=$(find . -maxdepth 1 -type f -printf "%f\n"|egrep '^headless-shell-[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+\.tar\.bz2'|egrep -v "($(join_by '|' $VERSIONS))"||:)
-if [ ! -z "$ARCHIVES" ]; then
-  (set -x;
-    rm -rf $ARCHIVES
-  )
-fi
-popd &> /dev/null
 
 # remove docker containers
 CONTAINERS=$(docker container ls \
